+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.2] - 2023-10-05
+
+### Changed
+
+- Updated `Dropdown` component
+  - Added `minHeight` and `maxHeight` props
+  - Updated style property from `minWidth` to `width`
+
+### Removed
+
+- Remove `grow` prop from `Dropdown` component
+
 ## [2.4.1] - 2023-10-05
 
 ### Added